backend/requirements.txt
anthropic>=0.96.0
google-genai>=1.70.0
google-auth>=2.35.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.0
httpx>=0.27.0
